Mark code that should be unreachable (#77)
Introduce a new [[noreturn]] unreachable() function that marks code as unreachable and throws a logic error if it is executed.
Fix build error due to Windows min/max macros.
Fix linker errors from referring to a non-constexpr extern from a constexpr.
Fix warnings about narrowing conversions.

+++ rmw_cyclonedds_cpp/src/serdata.hpp
@@ -43,15 +43,15 @@ struct sertopic_rmw : ddsi_sertopic
 class serdata_rmw : public ddsi_serdata
 {
 protected:
-  uint32_t m_size {0};
+  size_t m_size {0};
   /* first two bytes of data is CDR encoding
      second two bytes are encoding options */
   std::unique_ptr<byte[]> m_data {nullptr};
 
 public:
   serdata_rmw(const ddsi_sertopic * topic, ddsi_serdata_kind kind);
-  void resize(uint32_t requested_size);
-  uint32_t size() const {return m_size;}
+  void resize(size_t requested_size);
+  size_t size() const {return m_size;}
   void * data() const {return m_data.get();}
 };
